// src/const/type/Character.enum.ts
export enum Character {
    banjo_and_kazooie = 'banjo_and_kazooie',
    bayonetta = 'bayonetta',
    bowser = 'bowser',
    bowser_jr = 'bowser_jr',
    byleth = 'byleth',
    captain_falcon = 'captain_falcon',
    charizard = 'charizard',
    chrom = 'chrom',
    cloud = 'cloud',
    corrin = 'corrin',
    daisy = 'daisy',
    dark_pit = 'dark_pit',
    dark_samus = 'dark_samus',
    diddy_kong = 'diddy_kong',
    donkey_kong = 'donkey_kong',
    dq_hero = 'dq_hero',
    dr_mario = 'dr_mario',
    duck_hunt = 'duck_hunt',
    falco = 'falco',
    fox = 'fox',
    ganondorf = 'ganondorf',
    incineroar = 'incineroar',
    greninja = 'greninja',
    ice_climbers = 'ice_climbers',
    ike = 'ike',
    inkling = 'inkling',
    isabelle = 'isabelle',
    ivysaur = 'ivysaur',
    jigglypuff = 'jigglypuff',
    joker = 'joker',
    kazuya = 'kazuya',
    ken = 'ken',
    king_dedede = 'king_dedede',
    king_k_rool = 'king_k_rool',
    kirby = 'kirby',
    link = 'link',
    little_mac = 'little_mac',
    lucario = 'lucario',
    lucas = 'lucas',
    lucina = 'lucina',
    luigi = 'luigi',
    mario = 'mario',
    marth = 'marth',
    mega_man = 'mega_man',
    meta_knight = 'meta_knight',
    mewtwo = 'mewtwo',
    mii_fighter = 'mii_fighter',
    minmin = 'minmin',
    mr_game_and_watch = 'mr_game_and_watch',
    mythra = 'mythra',
    ness = 'ness',
    olimar = 'olimar',
    palutena = 'palutena',
    peach = 'peach',
    pichu = 'pichu',
    pikachu = 'pikachu',
    piranha_plant = 'piranha_plant',
    pit = 'pit',
    pyra = 'pyra',
    richter = 'richter',
    ridley = 'ridley',
    rob = 'rob',
    robin = 'robin',
    rosalina_and_luma = 'rosalina_and_luma',
    roy = 'roy',
    ryu = 'ryu',
    samus = 'samus',
    sephiroth = 'sephiroth',
    sheik = 'sheik',
    shulk = 'shulk',
    simon = 'simon',
    snake = 'snake',
    sonic = 'sonic',
    sora = 'sora',
    squirtle = 'squirtle',
    steve = 'steve',
    terry = 'terry',
    toon_link = 'toon_link',
    villager = 'villager',
    wario = 'wario',
    wii_fit_trainer = 'wii_fit_trainer',
    wolf = 'wolf',
    yoshi = 'yoshi',
    young_link = 'young_link',
    zelda = 'zelda',
    zero_suit_samus = 'zero_suit_samus',
}