// src/const/type/Series.enum.ts
export enum Series {
    animalcrossing_series = 'animalcrossing_series',
    donkeykong_series = 'donkeykong_series',
    finalfantasy_series = 'finalfantasy_series',
    kirby_series = 'kirby_series',
    mario_series = 'mario_series',
    metroid_series = 'metroid_series',
    mother_series = 'mother_series',
    palutena_series = 'palutena_series',
    pokemon_series = 'pokemon_series',
    starfox_series = 'starfox_series',
    streetfighter_series = 'streetfighter_series',
    xenoblade_series = 'xenoblade_series',
    zelda_series = 'zelda_series',
}